requirements.txt
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
install==1.3.3
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
requests==2.24.0
rsa==4.6
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
